Make it possible to easily plug in another stock prices API keeping up with the open-closed principle
