+++ src/test/java/com/melon/portfoliomanager/services/StockPricesServiceTests.java
@@ -3,6 +3,7 @@ package com.melon.portfoliomanager.services;
 import com.melon.portfoliomanager.dtos.responses.StockPricesDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
@@ -11,7 +12,6 @@ import org.springframework.http.ResponseEntity;
 
 import java.util.Map;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @SpringBootTest
@@ -19,11 +19,13 @@ public class StockPricesServiceTests {
     @org.springframework.boot.test.context.TestConfiguration
     public static class TestConfiguration {
         @Bean
+        @Qualifier("stock-prices-mock-api")
         @Primary
         public StockPricesHttpService stockPricesHttpServiceMock() {
             StockPricesHttpService stockPricesHttpService = mock(StockPricesHttpService.class);
             StockPricesDTO stockPricesDTO = new StockPricesDTO(Map.of("IBM", 1.23, "FB", 2.34));
-            doReturn(ResponseEntity.status(HttpStatus.OK).body(stockPricesDTO)).when(stockPricesHttpService).getStockPrices(any());
+            doReturn(ResponseEntity.status(HttpStatus.OK).body(stockPricesDTO))
+                    .when(stockPricesHttpService).getStockPrices();
             return stockPricesHttpService;
         }
     }
@@ -32,7 +34,7 @@ public class StockPricesServiceTests {
     private StockPricesHttpService stockPricesHttpService;
 
     @Test
-    public void testNotifyUserAboutExtremePriceChangesWillSucceed() {
-        verify(stockPricesHttpService, times(1)).getStockPrices(any());
+    public void testFetchCompanyStocksWillSucceed() {
+        verify(stockPricesHttpService, times(1)).getStockPrices();
     }
 }
